Moved everything into new engine and ui subpackages
