+++ pyro/engine/__init__.py
@@ -1,0 +1,4 @@
+from actor import Actor
+from engine import GameEngine, GameResult, Event, EventType
+from hero import Hero
+from monster import Monster
