+++ pyro/ui/screen.py
@@ -1,0 +1,11 @@
+
+
+class Screen:
+    def __init__(self):
+        pass
+
+    def update(self):
+        pass
+
+    def render(self):
+        pass
